@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { IAdminUser, ICreateUpdateAdminUser } from './admin-user.interface';
 import { of, from } from 'rxjs';
 import { mergeMap } from 'rxjs/operators';
-import { MatDialog, MatSnackBar } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { CreateUpdateUserComponent } from './create-update-user/create-update-user.component';
 import { catchError } from 'rxjs/operators';
 import { ModalService } from '../../shared/services/modal.service';
