@@ -4,7 +4,9 @@ import { Group, Student, UserAnswers } from 'src/app/shared/entity.interface';
 import { Test, Results } from '../entity.interface';
 import { ResultsService } from './results.service';
 import { ModalService } from '../../shared/services/modal.service';
-import { MatTable, MatTableDataSource, MatDialog, MatSort } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSort } from '@angular/material/sort';
+import { MatTable, MatTableDataSource } from '@angular/material/table';
 import { ResultRaitingQuestionComponent } from './result-raiting-question/result-raiting-question.component';
 import { ResultDetailComponent } from './result-detail/result-detail.component';
 import { forkJoin } from 'rxjs';
