@@ -2,8 +2,14 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { Speciality } from '.././../entity.interface';
 import { ApiService } from '../../../shared/services/api.service';
 import { DialogFormComponent } from '../dialog-form/dialog-form.component';
-import { MatDialog, MatTableDataSource, MatTable, MatPaginator, MatSnackBar } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatTableDataSource, MatTable } from '@angular/material/table';
 import { ModalService } from '../../../shared/services/modal.service';
+import { Store, select } from '@ngrx/store';
+import { AdminState } from '../../store/MainReducer';
+import { allSpecialitiesLoaded } from '../../store/speciality/speciality-actions';
 @Component({
   selector: 'app-speciality-list',
   templateUrl: './speciality-list.component.html',
@@ -15,14 +21,15 @@ export class SpecialityListComponent implements OnInit {
   public displayedColumns: string[] = ['code', 'name', 'buttons'];
   public dataSource = new MatTableDataSource<Speciality>();
 
-  @ViewChild('table', { static: false }) table: MatTable<Element>;
+  @ViewChild('table') table: MatTable<Element>;
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
 
   constructor(
     private apiService: ApiService,
     private dialog: MatDialog,
     private snackBar: MatSnackBar,
-    private modalService: ModalService) { }
+    private modalService: ModalService,
+    private store: Store<AdminState>) { }
 
   ngOnInit() {
     this.getSpeciality();
@@ -31,7 +38,10 @@ export class SpecialityListComponent implements OnInit {
     this.dataSource.paginator = this.paginator;
   }
   getSpeciality(): any {
-    this.apiService.getEntity('Speciality').subscribe((data: Speciality[]) => this.dataSource.data = data);
+    this.apiService.getEntity('Speciality').subscribe((data: Speciality[]) =>{
+       this.dataSource.data = data;
+      this.store.dispatch(allSpecialitiesLoaded({specialities: data}));
+      });
   }
 
   openConfirmDialog(speciality: Speciality) {
