@@ -6,13 +6,20 @@ import { HttpClientModule, HTTP_INTERCEPTORS, HttpClient } from '@angular/common
 import { ApiHttpInterceptor } from './http.interceptor';
 import { SharedModule } from './shared/shared.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatFormFieldModule, MatInputModule, MatPaginatorIntl } from '@angular/material';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatPaginatorIntl } from '@angular/material/paginator';
 import { NotFoundComponent } from './not-found/not-found.component';
 import { SpinnerInterceptor } from './shared/spinner/spinner.interceptor';
 import { TranslateModule, TranslateLoader, TranslateService } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { getMatPaginatorUkr } from './shared/mat-paginator-config/mat-pagination-intl';
 import { AngularWebStorageModule } from 'angular-web-storage';
+import { StoreModule } from '@ngrx/store';
+import { reducers, metaReducers, getInitialState } from './reducers';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { environment } from '../environments/environment';
+import { EffectsModule } from '@ngrx/effects';
 
 @NgModule({
   declarations: [
@@ -35,6 +42,16 @@ import { AngularWebStorageModule } from 'angular-web-storage';
     MatFormFieldModule,
     MatInputModule,
     AngularWebStorageModule,
+    StoreModule.forRoot(reducers, {
+      initialState: getInitialState,
+      metaReducers,
+      runtimeChecks: {
+        strictStateImmutability: true,
+        strictActionImmutability: true
+      }
+    }),
+    StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: environment.production }),
+    EffectsModule.forRoot([])
   ],
   providers: [
     {provide: HTTP_INTERCEPTORS, useClass: ApiHttpInterceptor, multi: true },
