@@ -1,6 +1,6 @@
 export const errorMapping = [
     {status: 0, message: 'interceptor.status0message'},
-    {status: 400, message: 'interceptor.status400message'},
+    // {status: 400, message: 'interceptor.status400message'},
     {status: 401, message: 'interceptor.status401message'},
     {status: 403, message: 'interceptor.status403message'},
     {status: 404, message: 'interceptor.status404message'},
