@@ -10,7 +10,10 @@ import { TestPlayerComponent } from './test-player/test-player.component';
 import { QuestionMenuItemComponent } from './question-menu-item/question-menu-item.component';
 import { QuestionAnswerComponent } from './question-answer/question-answer.component';
 import { TestPlayerService } from './test-player.service';
-import {MatButtonModule, MatIconModule, MatTableModule, MatTabsModule} from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTableModule } from '@angular/material/table';
+import { MatTabsModule } from '@angular/material/tabs';
 import { StudentInfoComponent } from './student-info/student-info.component';
 import {CountdownModule} from 'ngx-countdown';
 import {StudentInfoService} from './student-info.service';
